Lift psu_powershell out of SignalR transport so it can be used across that and gRPC.
